+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- Range assignments with images now resize the image before pasting
 
 ## [0.1.0] - 2025-07-21